Update build.gradle
This should fix a bug I found when setting up a dev environment for an addon that uses Artifacts API. It was putting two copies of CC in the dev environment, one under the maven group "io.github.OnyxStudios.Cardinal-Components-API" and one under "io.github.onyxstudios.Cardinal-Components-API", and since the CC README says to use the lowercase version, the latter would likely offer better compatibility. Note that without the temporary fix I used in my buildscript, the project could build as normal, but the launch profile would not run correctly.